+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 * Check if Polylang is installed
 * Check if query vars are set before testing them
+* Elementor ads
+* Hustle ads
 
 ### 1.0.9: October 11th, 2018
 * YoastSEO: do not copy title and metadesc