+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Elementor ads
 * Hustle ads
 * Add order count in menu (for shop_order_manager role)
+* Remove ACF ads
 
 ### 1.0.9: October 11th, 2018
 * YoastSEO: do not copy title and metadesc